fix: dispatch click away events to the right view
To choose the view to send the events to, the clicked element was being used.
If the clicked element belongs, for instance, to a sticky live view, but the
elements with the phx-click-away binding belong to the main live view (which
is the live view that knows about those elements), the events were being sent
to the sticky live view instead producing JS errors in the browser's console,
like `"no component found matching phx-target of 10 undefined"`, and sometimes
crashing the sticky live view too (because the event reached a live component
inside the sticky live view).